% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_log.R
\name{parse_log}
\alias{parse_log}
\title{Parse log file}
\usage{
parse_log(logfile, ...)
}
\arguments{
\item{logfile}{...}

\item{...}{arguments passed to parse operation}
}
\value{
parsed log
}
\description{
Parse log file
}

