% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_log_dir.R
\name{parse_log_dir}
\alias{parse_log_dir}
\title{Parse a directory of log files}
\usage{
parse_log_dir(log_dir, ...)
}
\arguments{
\item{log_dir}{...}

\item{...}{arguments passed to parse operation}
}
\value{
parsed log
}
\description{
Parse a directory of log files
}

